\name{numerovmpreal-package}
\alias{numerovmpreal-package}
\alias{numerovmpreal}
\docType{package}
\title{
\packageTitle{numerovmpreal}
}
\description{
\packageDescription{numerovmpreal}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{numerovmpreal}
\packageIndices{numerovmpreal}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{numerovmpreal}

Maintainer: \packageMaintainer{numerovmpreal}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
